% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lukesky-package.R
\docType{package}
\name{lukesky-package}
\alias{lukesky}
\alias{lukesky-package}
\title{lukesky: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
