% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_error.R
\name{sd_error}
\alias{sd_error}
\title{standard error function}
\usage{
sd_error(x)
}
\arguments{
\item{x}{}
}
\value{
a numeric value
}
\description{
A function that allows you to put a vector of numric values and then calculates standard error based on sd and sqrt
}
